Move transtition_probs function to new module.
